PDI-1632: Add Testing Framework to PingCTL
- Add SSO resource unit testing
- Update pingone_application_flow_policy.go to use the policy name
instead of ID for ResourceName in the ImportBlock.
- Update pingone_application_resource_grant.go to use the resource
name instead of ID for ResourceName in the ImportBlock.
- Update pingone_application_role_assignment.go to use an index number
in ResourceName for the ImportBlock to avoid ResourceName duplication.
- Update pingone_application_sign_on_policy_assignment.go to use the
sign on policy name instead of ID for ResourceName in the ImportBlock.
- Update pingone_group_role_assignment.go to use the role name instead
of ID for ResourceName in the Importblock, as well as an index number
to avoid ResourceName duplication.
- Update pingone_sign_on_action.go to use the action type instead
of ID for ResourceName in the ImportBlock.
- Update helper test function ValidateImportBlocks to check for
resource name and id duplication first after rework to use import
block map instead of list for expected import blocks.
- Update helper test function to check if the retrieved expected
import block is empty, meaning there was no match with the currently
compared generated import block.