PDI-1652: Prevent creation of import blocks for immutable PingOne config
- Update export.go to supply the api client ID to connectors. This is
used by the application_secret resource.
- Update exportable_resource.go to support the api client id in
the SDKClientInfo struct.
- Update pingone_application_secret.go and tests to check if the
export app ID is the same as the worker ID used to configure the
API client. In this case, we expect a 403 response.
- Update pingone_resource_attribute.go to not export any attributes
that are CORE types, with the exception of 'sub' for CUSTOM resources.
- Update pingone_resource_scope.go to ensure only non-OpenID and
non-PingOneAPI scopes are exported for this resource.
- Update pingone_resource_scope_openid.go similarly.
- Update pingone_resource_scope_pingone_api.go to check scope names
to be of the allowed patterns:
p1:read:user, p1:update:user, p1:read:user:{suffix}, or p1:update:user:{suffix}.